Moves utility functions and checks to utils/pavlov
